move directive to core-ui
