+++ src/app/core-ui/directive/common.directives.ts
