+++ src/app/core-ui/main/main-view.module.ts
@@ -4,7 +4,7 @@ import { RouterModule } from '@angular/router';
 import { MatIconModule } from '@angular/material';
 
 import { MaterialModule } from '../material/material.module';
-import { DirectiveModule } from '../../directive/directive.module';
+import { DirectiveModule } from '../directive/directive.module';
 
 import { MainViewComponent } from './main-view.component';
 import { StatusComponent } from './status/status.component';
